+++ GridMvc.Site/Models/Grids/OrdersAjaxPagingGrid.cs
@@ -27,7 +27,7 @@ namespace GridMvc.Site.Models.Grids
         public int CurrentPage { get; set; }
 
         public int? ItemsCountOverwrite { get; set; }
-        public bool? CustomPaging { get; set; }
+        public bool CustomPaging { get; set; }
 
         public string TemplateName
         {